requirements.txt
anthropic>=0.3.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
pandas>=1.3.0
asana>=0.10.0